Fix: Weight Type Int zu Float und neues GoalEntity in den anderen Klassen hinzugefügt

+++ Spring/src/main/java/de/sep/calocalendar/entities/Goal.java
@@ -17,6 +17,6 @@ public class Goal {
     @Id
     private Long id;
 
-    private int goalWeight;
+    private float goalWeight;
     private int goalSpeed;
 }
